chore: moved RoleRepository
